Feature/Asp .Net Authentication with Identity Cookies

+++ LiteStreaming.DataSeeders/Implementations/Administrative/DataSeederAdministrative.cs
@@ -12,6 +12,12 @@ public class DataSeederAdministrative : IDataSeeder
     {
         try
         {
+            if (_context.Database.CanConnect())
+            {
+                Console.WriteLine("Banco de dados já existe. Não será recriado.");
+                return;
+            }
+
             _context.Database.EnsureDeleted();
             _context.Database.EnsureCreated();
